Split console completion rendering (#4170)
