+++ changelog.d/unreleased/4170.internal.md
@@ -1,0 +1,19 @@
+---
+category: internal
+issues:
+  - 4170
+affected:
+  - src/CodeIndex/Cli/CliCommandCatalog.cs
+  - src/CodeIndex/Cli/ConsoleCompletionRenderer.cs
+  - src/CodeIndex/Cli/ConsoleUi.cs
+  - tests/CodeIndex.Tests/CliFlagSchemaTests.cs
+  - tests/CodeIndex.Tests/ConsoleUiTests.cs
+---
+
+## English
+
+- **Console completion rendering now has focused ownership (#4170)** — bash, zsh, fish, and PowerShell completion generation now lives in `ConsoleCompletionRenderer`, with shared command names in `CliCommandCatalog` and snapshot-style shell format checks pointed at the completion renderer directly.
+
+## 日本語
+
+- **コンソール補完 rendering の所有を分離しました (#4170)** — bash、zsh、fish、PowerShell の completion 生成は `ConsoleCompletionRenderer` に移り、共有 command 名は `CliCommandCatalog` に分離し、snapshot-style の shell format 検証は completion renderer を直接確認するようになりました。
