+++ src/CodeIndex/Cli/CliCommandCatalog.cs
@@ -1,0 +1,11 @@
+namespace CodeIndex.Cli;
+
+internal static class CliCommandCatalog
+{
+    internal static readonly string[] Commands =
+    [
+        "index", "backfill-fold", "optimize", "search", "recipes", "audit", "definition", "goto", "references", "callers", "callees",
+        "symbols", "files", "find", "excerpt", "map", "inspect", "outline", "status", "validate-config",
+        "validate", "deps", "impact", "unused", "hotspots", "suggestions", "languages", "batch", "mcp", "completions", "db", "vacuum", "report", "license", "upgrade",
+    ];
+}
